chore: split out endpoints
