+++ backend/src/routes/index.ts
@@ -6,11 +6,13 @@ import { ServiceContext } from 'twilio/lib/rest/verify/v2/service'
 import type { Database } from '@/db/drizzle'
 
 import { getClaimRoute } from './claim'
+import { createFunkitStripeCheckout } from './createFunkitStripeCheckout'
 import { getExecuteFromOutsideRoute } from './executeFromOutside'
-import { createFunkitStripeCheckout, getFunkitStripeCheckoutQuote, getFunkitStripeCheckoutStatus } from './funkit'
 import { getGenerateClaimLinkRoute } from './generateClaimLink'
 import { getBalanceRoute } from './getBalance'
 import { getCurrentExpenseRoute } from './getCurrentExpense'
+import { getFunkitStripeCheckoutQuote } from './getFunkitStripeCheckoutQuote'
+import { getFunkitStripeCheckoutStatus } from './getFunkitStripeCheckoutStatus'
 import { getHistoricalBalanceRoute } from './getHistoricalBalance'
 import { getLimitRoute } from './getLimit'
 import { getOtp } from './getOtp'
